chore: Move some of the functions for multi-upload into another file
